Fix: URL normalization not as in RFC 3986
Net_URL2 did not properly address URL normilization as
outlined in RFC 3986 Section 6.

The normalize() method had two missing rungs in the
comparison ladder:

Percent-Encoding Normalization (6.2.2.2.) was not
addressed with the result that percent-encoded octets
that do not require percent-encoding were not normalized
to their unreserved character.

Scheme-Based Normalization (6.2.3.) was broken for the
default port in form of an empty string. This flaw was
the result of a missing string check.

These two issues are adressed with these fixes.

Steps done:

- Test extended with examples from RFC 3986 Section 6
- Fixed percent encoding in class Net_Url2 _normalizeCallback() method
- Fixed missing type-check in class Net_Url2 normalize() method
- Improved by using an array instead of foreach'ing over it
- Typo fixed; RFC number is 3986 (previously wrongly 3886)
- WS Fixed