Fixed #19684: Redirects containing spaces do not work
A.k.a Recover from unencoded characters.

Net_URL2 was not able to recover from data within query and path
that was neither within the reserved, unreserved space nor
characters that have a reserved meaning (like %).

The issue was reported as Pear Request #19684 on 2012-10-30 21:52 UTC.

The solution is to encode those characters:

    [\x0-\x20\x22\x3C\x3E\x7F-\xFF]

Which is similar to what modern browsers do when an URL is put into
their address-bar.

Steps done:

- Add Test 19684
- Fixed Pear Request #19684 in class Net_Url2 parseUrl() method