Fix: Setting an empty userinfo results in commercial at-sign ("@")
Net_URL2 did not properly address handling a zero-
length string for the user-information (userinfo in
short).

According to RFC 3986 the commercial at-sign ("@")
should only be used if the user information is
present.

An empty userinfo (zero-length-string) is not re-
presenting such a userinfo. Therefore the user-
info is not present.

The original report by Ross Perkins was pointing
to a regression introduced in 13e769b stringy-
fying a bool.

Steps done:

- Test extended setting userinfo to zero-length
  string by adding testSetEmptyUserinfo() method
- Fixed ::$_userinfo string handling in Net_URL2
  getAuthority() method

References:

- https://pear.php.net/bugs/bug.php?id=20013 (#20013)
- https://github.com/pear/Net_URL2/pull/3 (#3)