Fixed #20157: normalize() flaw with "0" as path
The special case path "0" (zero string) triggered a flaw in
normalize() replacing it with "/". This has been fixed.

Additionally, if the path was not prefixed with "/"
while there was authority, it was not automatically prefixed
with "/". This has been fixed.

Note: This is not part of URL normalization (and also missing
at other places), however convenient to fix at this time at
this place.

Steps done:

- Add Test 20157
- Fixed Pear Bug #20157: normalize() flaw with '0' as path