Reviewed: Fix CS, license text quirks in docs
The docs folder did not comply with PEAR Coding Standards.

The code has been fixed. The invalid header of the file has been
condensed.

Licensing terms have been preserved as a single file.

The text had little difference with the online reference w/o
semantical differences so that it has been streamlined to it.

The following changes were done:

1.) The names of the authors may not be used to endorse -> Neither the
name of the Richard Heyes nor the names of his contributors may be used to endorse
2.) COPYRIGHT OWNER -> COPYRIGHT HOLDER