New: NetURL2::removeDotSegments() emits a warning on loop limit reach
NetURL2::removeDotSegments() contains a limit of iterations done. When
this limit is reached, a warning has been added to inform about that.

Currently the limit is 100 iterations. This will be raised after
more tests for that function have been added remove dot segments in
more than 100 path segments.

This change makes also a flaw with a type check visible that was intro-
duced in fb91fb772cfe4008bac44bc28ae547be04f96fdb .