Fix: NetURL2::removeDotSegments() substr failure to comply
PHPs substr() can not obtain a zero-length string from the end of the
string, instead it handles this as a failure as the input-parameter of
the start offset seems too high.

As  NetURL2::removeDotSegments() uses more strict type-checks since
fb91fb7 this resulted in a flaw in the internal order of
removeDotSegments main loop due to a the typecheck change introduced.

Previously FALSE returned by substr() did exit the loop, this was
not the case any longer. This has been fixed.

This flaw was made visible by adding the warning message on loop limit
reach.