Fixed #20423: URI with IPv6 or IPvFuture not parsed
Net_URL2 did not accept IP-literal for host which caused a major problem
with URIs containing these.

This has been fixed by improving authority parsing to detect the host
within an authority in a more lax fashion by looking for it only by its
surrounding userinfo "@" and ":" port.

Additionally there was a flaw that the authority parser did not accept the
optional port-number which has been fixed as well.