Fixed #20418: Incorrect normalization of URI with missing authority
Net_URL2::normalize() did not skip unset host and path during normalization
which resulted in setting the according fields to an empty string in error.

This has been fixed by treating the FALSE case exactly like it was already
done for the userinfo field which was reported earlier in #20385 preventing
the unintended type conversion.